fix: make moderation delta arithmetic explicit integers
Agent-Logs-Url: https://github.com/nickmsk9/litetracker/sessions/466fb532-993f-49ea-8eea-999c8cce8ed8

+++ ajax/profile.php
@@ -387,11 +387,11 @@ if ($action === 'moderate_profile') {
 	$uploadedMb = (int) ($_POST['uploaded_mb'] ?? 0);
 	$downloadedMb = (int) ($_POST['downloaded_mb'] ?? 0);
 	if ($uploadedMb !== 0) {
-		$delta = $uploadedMb * 1024 * 1024;
+		$delta = (int) ($uploadedMb * 1024 * 1024);
 		$updates[] = "uploaded=GREATEST(uploaded + (".$delta."), 0)";
 	}
 	if ($downloadedMb !== 0) {
-		$delta = $downloadedMb * 1024 * 1024;
+		$delta = (int) ($downloadedMb * 1024 * 1024);
 		$updates[] = "downloaded=GREATEST(downloaded + (".$delta."), 0)";
 	}
